SHS module service
